+++ backend/canisters/group/impl/src/updates/add_reaction.rs
@@ -30,6 +30,7 @@ fn add_reaction_impl(args: Args, state: &mut RuntimeState) -> Response {
             args.message_id,
             args.reaction.clone(),
             now,
+            &mut state.data.event_sink_client,
         ) {
             AddRemoveReactionResult::Success => {
                 handle_activity_notification(state);