+++ backend/canisters/user/impl/src/updates/tip_message.rs
@@ -148,7 +148,10 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<(PrepareResult, Timestam
 
 fn tip_direct_chat_message(args: TipMessageArgs, decimals: u8, state: &mut RuntimeState) -> Response {
     if let Some(chat) = state.data.direct_chats.get_mut(&args.recipient.into()) {
-        match chat.events.tip_message(args.clone(), EventIndex::default()) {
+        match chat
+            .events
+            .tip_message(args.clone(), EventIndex::default(), Some(&mut state.data.event_sink_client))
+        {
             TipMessageResult::Success => {
                 state.push_user_canister_event(
                     args.recipient.into(),